requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
CacheControl==0.13.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dep-logic==0.0.4
distlib==0.3.8
distro==1.9.0
docstring-parser==0.15
exceptiongroup==1.2.0
filelock==3.13.1
findpython==0.4.1
Flask==3.0.2
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.0
installer==0.7.0
instructor==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.3
langsmith==0.1.22
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.5
openai==1.14.3
orjson==3.9.15
packaging==23.2
pdm==2.11.1
platformdirs==4.1.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyproject_hooks==1.0.0
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
resolvelib==1.0.1
rich==13.7.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.10.0
unearth==0.12.1
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
